feat: add math recognize
