+++ .pre-commit-config.yaml
@@ -5,7 +5,7 @@ repos:
     rev: 5.0.4
     hooks:
       - id: flake8
-        args: ["--max-line-length=200", "--ignore=E131,E125,W503,W504,E203,E231"]
+        args: ["--max-line-length=500", "--ignore=E131,E125,W503,W504,E203,E231"]
   - repo: https://github.com/PyCQA/isort
     rev: 5.11.5
     hooks:
